[sg noup] net: lwm2m: Automatic send mechanism
Documentation for this feature can be found in our sg-zephyr-firmware
repo.

It seems like nobody besides us uses LwM2M send, otherwise we would not
have to build something on our own.

This commit noup because upstreaming it would require substantial
software design and discussion work upstream.

The command to sync all IPSOs is intended as a fall-back solution for
any kind of issues in the field.

Issue: R2D2-10191
Issue: R2D2-10635
Issue: R2D2-10946
Issue: R2D2-11646

Auto-send is easier to maintain if we move the implementation to
`sg-zephyr-firmware` and keep only the nesseccary function calls and
modifications in the zephyr fork. Keeping it in-tree would make upstreaming
easier, but this is not planned any time soon.

(cherry picked from commit 9b561e20b2130a28fa9249e3415a2e3ce55d4cb6)

+++ subsys/net/lib/lwm2m/lwm2m_observation.h
@@ -62,7 +62,7 @@ struct lwm2m_obj_path_list {
 /* Initialize path list */
 void lwm2m_engine_path_list_init(sys_slist_t *lwm2m_path_list, sys_slist_t *lwm2m_free_list,
 				 struct lwm2m_obj_path_list path_object_buf[],
-				 uint8_t path_object_size);
+				 uint16_t path_object_size);
 /**
  * Add new path to the list while maintaining hierarchical sort order
  *
